Scalar_Lab/requirements.txt
ollama
pyautogen>=0.2.0
agentlightning
MPh>=1.3.1
pyaedt
numpy
pandas
scipy
matplotlib